[REFACTOR] 마이페이지 조회 코드 리팩토링
