+++ src/main/java/com/pharmquest/pharmquest/domain/mypage/service/MyPageServiceImpl.java
@@ -1,29 +1,29 @@
 package com.pharmquest.pharmquest.domain.mypage.service;
 
-import com.pharmquest.pharmquest.domain.mypage.converter.MyPageConverter;
+import com.pharmquest.pharmquest.domain.mypage.web.dto.MyPageResponseDTO;
+import com.pharmquest.pharmquest.domain.pharmacy.data.enums.PharmacyCountry;
+import com.pharmquest.pharmquest.domain.pharmacy.service.PharmacyDetailsService;
 import com.pharmquest.pharmquest.domain.supplements.data.Supplements;
 import com.pharmquest.pharmquest.domain.supplements.data.mapping.SupplementsScrap;
-import com.pharmquest.pharmquest.domain.supplements.repository.SupplementsRepository;
 import com.pharmquest.pharmquest.domain.supplements.repository.SupplementsScrapRepository;
 import com.pharmquest.pharmquest.domain.user.data.User;
-import com.pharmquest.pharmquest.domain.user.repository.UserRepository;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.NoSuchElementException;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 @Slf4j
 public class MyPageServiceImpl implements MyPageService {
 
-    private final UserRepository userRepository;
+    private final PharmacyDetailsService pharmacyDetailsService;
     private final SupplementsScrapRepository supplementsScrapRepository;
+
     @Override
     public List<Supplements> getScrapSupplements(Long userId) {
 
@@ -35,5 +35,18 @@ public class MyPageServiceImpl implements MyPageService {
         return supplementsScrapList.stream()
                 .map(SupplementsScrap::getSupplements) // SupplementsScrap에서 supplements 값을 추출
                 .toList();
+    }
+
+    @Override
+    public List<MyPageResponseDTO.PharmacyDto> getScrapPharmacies(User user, String country) {
+
+        List<String> pharmacyPlaceIdList = user.getPharmacyScraps();
+        String findingCountryName = PharmacyCountry.getCountryByName(country).getGoogleName(); // Query String으로 입력받은 국가의 google에 등록된 이름으로 변경
+
+        return pharmacyPlaceIdList.stream()
+                .map(pharmacyDetailsService::getPharmacyDtoByPlaceId)
+                .filter(pharmacyDto -> findingCountryName.equals(pharmacyDto.getCountry()) || findingCountryName.equals("all"))
+                .toList();
+
     }
 }
