--- src/main/java/com/pharmquest/pharmquest/domain/mypage/service/ScrapService.java
@@ -1,10 +1,0 @@
-package com.pharmquest.pharmquest.domain.mypage.service;
-
-import com.pharmquest.pharmquest.domain.mypage.web.dto.ScrapResponseDTO;
-import com.pharmquest.pharmquest.domain.user.data.User;
-
-import java.util.List;
-
-public interface ScrapService {
-    public List<ScrapResponseDTO.PharmacyDto> getPharmacies(User user, String country);
-}
