+++ src/main/java/com/pharmquest/pharmquest/domain/pharmacy/service/PharmacyDetailsService.java
@@ -1,10 +1,10 @@
 package com.pharmquest.pharmquest.domain.pharmacy.service;
 
-import com.pharmquest.pharmquest.domain.mypage.web.dto.ScrapResponseDTO;
+import com.pharmquest.pharmquest.domain.mypage.web.dto.MyPageResponseDTO;
 
 
 public interface PharmacyDetailsService {
     public Boolean isPharmacyByPlaceId(String placeId);
 
-    public ScrapResponseDTO.PharmacyDto getPharmacyDtoByPlaceId(String placeId);
+    public MyPageResponseDTO.PharmacyDto getPharmacyDtoByPlaceId(String placeId);
 }
