+++ src/main/java/com/pharmquest/pharmquest/domain/user/data/User.java
@@ -4,7 +4,6 @@ import com.pharmquest.pharmquest.domain.mypage.data.MedicineScrap;
 import com.pharmquest.pharmquest.domain.mypage.data.SupplementScrap;
 import com.pharmquest.pharmquest.domain.mypage.data.PostScrap;
 import com.pharmquest.pharmquest.domain.pharmacy.converter.PlaceIdConverter;
-import com.pharmquest.pharmquest.domain.pharmacy.converter.ListStringConverter;
 import com.pharmquest.pharmquest.domain.post.data.mapping.Comment;
 import com.pharmquest.pharmquest.domain.post.data.mapping.PostLike;
 import com.pharmquest.pharmquest.domain.post.data.mapping.PostReport;
